requirements.txt
pillow
matplotlib
numpy
keras
tensorflow
scipy
